SHOP-50: Mail sending setup + return optional when searching for account with email

+++ shopper-app/src/main/java/com/qprogramming/shopper/app/items/ListItemService.java
@@ -127,7 +127,7 @@ public class ListItemService {
     }
 
     private Product updateCategoryScore(Category updatedCategory, Product product) {
-        Integer categoryScore = product.getCategoryScore().getOrDefault(updatedCategory, 0);
+        Long categoryScore = product.getCategoryScore().getOrDefault(updatedCategory, 0L);
         product.getCategoryScore().put(updatedCategory, ++categoryScore);
         return _productRepository.save(product);
     }
